Move high-level event logic to call sites
