+++ core/lib/multivm/src/versions/vm_boojum_integration/vm.rs
@@ -1,6 +1,5 @@
 use circuit_sequencer_api_1_4_0::sort_storage_access::sort_storage_access_queries;
 use zksync_types::{
-    event::extract_l2tol1logs_from_l1_messenger,
     l2_to_l1_log::{SystemL2ToL1Log, UserL2ToL1Log},
     Transaction,
 };
@@ -14,6 +13,7 @@ use crate::{
         VmExecutionResultAndLogs, VmFactory, VmInterface, VmInterfaceHistoryEnabled,
         VmMemoryMetrics,
     },
+    utils::events::extract_l2tol1logs_from_l1_messenger,
     vm_boojum_integration::{
         bootloader_state::BootloaderState,
         old_vm::events::merge_events,
