+++ core/node/commitment_generator/src/tests/mod.rs
@@ -7,7 +7,7 @@ use zksync_dal::Connection;
 use zksync_node_genesis::{insert_genesis_batch, GenesisParams};
 use zksync_node_test_utils::{create_l1_batch, create_l2_block};
 use zksync_types::{
-    block::L1BatchTreeData, zk_evm_types::LogQuery, AccountTreeId, Address, StorageLog,
+    block::L1BatchTreeData, zk_evm_types::LogQuery, AccountTreeId, Address, StorageLog, VmEvent,
 };
 
 use super::*;
@@ -298,4 +298,26 @@ async fn commitment_generator_with_tree_emulation() {
     tree_emulator_handle.await.unwrap();
     stop_sender.send_replace(true);
     generator_handle.await.unwrap().unwrap();
+}
+
+#[test]
+fn test_convert_vm_events_to_log_queries() {
+    let cases: Vec<serde_json::Value> = vec![
+        serde_json::from_str(include_str!(
+            "./test_vectors/event_with_1_topic_and_long_value.json"
+        ))
+        .unwrap(),
+        serde_json::from_str(include_str!("./test_vectors/event_with_2_topics.json")).unwrap(),
+        serde_json::from_str(include_str!("./test_vectors/event_with_3_topics.json")).unwrap(),
+        serde_json::from_str(include_str!("./test_vectors/event_with_4_topics.json")).unwrap(),
+        serde_json::from_str(include_str!("./test_vectors/event_with_value_len_1.json")).unwrap(),
+    ];
+
+    for case in cases {
+        let event: VmEvent = serde_json::from_value(case["event"].clone()).unwrap();
+        let expected_list: Vec<LogQuery> = serde_json::from_value(case["list"].clone()).unwrap();
+
+        let actual_list = convert_vm_events_to_log_queries(&[event]);
+        assert_eq!(actual_list, expected_list);
+    }
 }
