+++ core/node/commitment_generator/src/tests/test_vectors/event_with_4_topics.json
