+++ core/node/commitment_generator/src/tests/test_vectors/event_with_value_len_1.json
