+++ core/node/state_keeper/src/updates/l2_block_updates.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use once_cell::sync::Lazy;
 use zksync_multivm::{
     interface::{
         CompressedBytecodeInfo, ExecutionResult, L2BlockEnv, TransactionExecutionResult,
@@ -7,9 +8,10 @@ use zksync_multivm::{
     },
     vm_latest::TransactionVmExt,
 };
+use zksync_system_constants::KNOWN_CODES_STORAGE_ADDRESS;
 use zksync_types::{
     block::{BlockGasCount, L2BlockHasher},
-    event::extract_bytecodes_marked_as_known,
+    ethabi,
     l2_to_l1_log::{SystemL2ToL1Log, UserL2ToL1Log},
     vm_trace::Call,
     L2BlockNumber, ProtocolVersionId, StorageLogWithPreviousValue, Transaction, VmEvent, H256,
@@ -17,6 +19,27 @@ use zksync_types::{
 use zksync_utils::bytecode::hash_bytecode;
 
 use crate::metrics::KEEPER_METRICS;
+
+/// Extracts all bytecodes marked as known on the system contracts.
+fn extract_bytecodes_marked_as_known(all_generated_events: &[VmEvent]) -> Vec<H256> {
+    static PUBLISHED_BYTECODE_SIGNATURE: Lazy<H256> = Lazy::new(|| {
+        ethabi::long_signature(
+            "MarkedAsKnown",
+            &[ethabi::ParamType::FixedBytes(32), ethabi::ParamType::Bool],
+        )
+    });
+
+    all_generated_events
+        .iter()
+        .filter(|event| {
+            // Filter events from the deployer contract that match the expected signature.
+            event.address == KNOWN_CODES_STORAGE_ADDRESS
+                && event.indexed_topics.len() == 3
+                && event.indexed_topics[0] == *PUBLISHED_BYTECODE_SIGNATURE
+        })
+        .map(|event| event.indexed_topics[1])
+        .collect()
+}
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct L2BlockUpdates {
